% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_SW_data.R
\name{analyze_sw_data}
\alias{analyze_sw_data}
\title{Title}
\usage{
analyze_sw_data(dat, outcome_type, method, estimand, time_varying_assumption)
}
\arguments{
\item{dat}{A dataframe containing the stepped wedge trial data.}

\item{outcome_type}{A character string; either "binary" or "continuous".}

\item{method}{A character string; either "mixed", for a mixed-effects model, or "GEE", for generalized estimating equations.}

\item{estimand}{A character string; either "TATE", for time-averaged treatment effect, or "LTE", for long-term treatment effect.}

\item{time_varying_assumption}{A character string; either "IT" for immediate treatment effect, or "ETI", for Exposure time indicator.}
}
\value{
A list with ___
}
\description{
Title
}
\examples{
# TO DO
}
