% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effect_curves.R
\name{as_table}
\alias{as_table}
\title{Create table of estimates}
\usage{
as_table(..., labels = NA)
}
\arguments{
\item{...}{One or more objects of class \code{"sw_analysis"} returned by
\code{\link{analyze}}.}

\item{labels}{A character vector of length equal to length(list(...))
representing curve labels}
}
\value{
A table of effect estimate values
}
\description{
Format estimates returned by \code{\link{analyze}} as a table
}
\examples{
# Load data
test_data <- load_data(time ="period", cluster_id = "cluster", individual_id = NULL,
treatment = "trt", outcome = "outcome_cont", data = sw_data_example)

\donttest{
IT_model <- analyze(dat = test_data, method = "mixed", estimand_type = "TATE",
estimand_time = c(1, 4), exp_time = "IT")
ETI_model <- analyze(dat = test_data, method = "mixed", estimand_type = "TATE",
estimand_time = c(1, 4), exp_time = "ETI")
NCS_model <- analyze(dat = test_data, method = "mixed", estimand_type = "TATE",
estimand_time = c(1, 4), exp_time = "NCS")
TEH_model <- analyze(dat = test_data, method = "mixed", estimand_type = "TATE",
estimand_time = c(1, 4), exp_time = "TEH")
ests_table <- as_table(IT_model, IT_model, NCS_model, TEH_model)
head(ests_table)
}
}
