% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sw_data.R
\name{load_sw_data}
\alias{load_sw_data}
\title{Load and format data object}
\usage{
load_sw_data(
  period,
  cluster_id,
  individual_id = NULL,
  treatment,
  covariates = NULL,
  outcome,
  time_type = "discrete",
  data
)
}
\arguments{
\item{period}{A character string; the name of the numeric variable representing the discrete time period of the observation.}

\item{cluster_id}{A character string; the name of the numeric variable identifying the cluster.}

\item{individual_id}{A character string (optional); the name of the numeric variable identifying the individual.}

\item{treatment}{A character string; the name of the binary variable indicating treatment. Accepts either integer (0/1) or Boolean (T/F) values.}

\item{covariates}{A character vector; the names of the covariate columns. Columns values should be either numeric, binary, or factors. Character columns will be converted into factors.}

\item{outcome}{A character string; the name of the numeric or binary variable indicating outcome. Accepts either numeric or Boolean (T/F) values.}

\item{time_type}{A character string describing how time is accounted for: 'discrete' (default) or 'continuous'}

\item{data}{A dataframe containing the stepped wedge trial data.}
}
\value{
An object of class \code{sw_data}
}
\description{
Load and format data object
}
\examples{
load_sw_data(period = "period", cluster_id = "id", individual_id = NULL,
treatment = "treatment", covariates = NULL, outcome = "y_bin",
data = geeCRT::sampleSWCRTLarge)
}
