% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot_design}
\alias{plot_design}
\title{Plot stepped wedge design}
\usage{
plot_design(dat)
}
\arguments{
\item{dat}{A dataframe containing the stepped wedge trial data.}
}
\value{
A list with the number of clusters, number of time points, a summary of the data, and a plot of the stepped wedge design.
}
\description{
Plot stepped wedge design
}
\examples{
# TO DO
}
