% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_sw_analysis.R
\name{print.sw_analysis}
\alias{print.sw_analysis}
\title{Title}
\usage{
\method{print}{sw_analysis}(x, ...)
}
\arguments{
\item{x}{A list returned by steppedwedge::analysis()}

\item{...}{Additional arguments}
}
\value{
A summary of the analysis object, including an estimate and 95% CI for the treatment effect, and messages on convergence.
}
\description{
Title
}
\examples{
# TO DO
}
