% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steppedwedge-package.R
\docType{package}
\name{steppedwedge-package}
\alias{steppedwedge}
\alias{steppedwedge-package}
\title{steppedwedge: Analyze Data from Stepped Wedge Cluster Randomized Trials}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provide various functions and tools to help fit models for estimating treatment effects in stepped wedge cluster randomized trials. Implements methods described in Kenny, Voldal, Xia, and Heagerty (2022) "Analysis of stepped wedge cluster randomized trials in the presence of a time-varying treatment effect", \doi{10.1002/sim.9511}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://avi-kenny.github.io/steppedwedge/}
  \item \url{https://github.com/Avi-Kenny/steppedwedge}
  \item Report bugs at \url{https://github.com/Avi-Kenny/steppedwedge/issues}
}

}
\author{
\strong{Maintainer}: Avi Kenny \email{avi.kenny@gmail.com} [copyright holder]

Authors:
\itemize{
  \item David Arthur \email{david.arthur291@duke.edu}
  \item Yongdong Ouyang \email{yongdong.ouyang@sickkids.ca}
}

}
\keyword{internal}
