% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sw_data_example}
\alias{sw_data_example}
\title{Example stepped wedge data}
\format{
## `sw_data_example`
A data frame with 2,063 rows and 5 columns:
\describe{
  \item{cluster}{Cluster id}
  \item{period}{Time period}
  \item{trt}{Treatment indicator}
  \item{outcome_bin}{Binary outcome}
  \item{outcome_cont}{Continuous outcome}
  ...
}
}
\usage{
sw_data_example
}
\description{
Data generated for the purpose of demonstrating the steppedwedge package
}
\keyword{datasets}
