% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sw_data_example_binom}
\alias{sw_data_example_binom}
\title{Example stepped wedge data with aggregrated outcomes}
\format{
## `sw_data_example_binom`
A data frame with 90 rows and 5 columns:
\describe{
  \item{cluster}{Cluster id}
  \item{period}{Time period}
  \item{trt}{Treatment indicator}
  \item{denominator}{Denominator (# trials)}
  \item{numerator}{Numerator (# successes)}
  ...
}
}
\usage{
sw_data_example_binom
}
\description{
A simulated stepped wedge dataset formatted with separate columns for
the number of successes (numerator) and number of trials (denominator).
}
\keyword{datasets}
