% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Title}
\usage{
analyze(
  dat,
  method = "mixed",
  estimand,
  exp_time = "IT",
  cal_time = "categorical",
  family,
  corstr = "exchangeable"
)
}
\arguments{
\item{dat}{A dataframe containing the stepped wedge trial data.}

\item{method}{A character string; either "mixed", for a mixed-effects model, or "GEE", for generalized estimating equations.}

\item{estimand}{A character string; either "TATE", for time-averaged treatment effect, or "LTE", for long-term treatment effect.}

\item{exp_time}{One of c("IT", "ETI", "NCS"); model for exposure time. "IT" encodes an immediate treatment model with a single treatment effect parameter. "ETI" is an exposure time indicator model, including one indicator variable for each exposure time point. "NCS" uses a natural cubic spline model for the exposure time trend.}

\item{cal_time}{One of c("categorical", "NCS", "linear", "none"); model for calendar time. "categorical" uses indicator variables for discrete time points, as in the Hussey and Hughes model. "NCS" uses a natural cubic spline, useful for datasets with continuous time. "linear" uses a single slope parameter. "none" assumes that there is no underlying calendar time trend.}

\item{family}{A family object; see documentation for `glm()`.}

\item{corstr}{A character string; see documentation for `geepack::geeglm()`.}
}
\value{
A list with ___
}
\description{
Title
}
\examples{
# TO DO
}
