% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Set advanced parameters for analysis of data from stepped wedge trials}
\usage{
params(
  offset = NULL,
  n_knots_exp = 4,
  n_knots_cal = 4,
  var_est = "model",
  var_est_type = "classic",
  return_ncs = F,
  re_correlated = F
)
}
\arguments{
\item{offset}{A linear predictor offset term; see docs for `lme4::lmer`.}

\item{n_knots_exp}{An integer; only relevant when exp_time="NCS". Specifies
the number of knots to use for exposure time, including boundary knots.
The spline basis includes an intercept, and the degree of the basis is
equal to the number of knots.}

\item{n_knots_cal}{An integer; only relevant when cal_time="NCS". Specifies
the number of knots to use for calendar time, including boundary knots.
The spline basis includes an intercept, and the degree of the basis is
equal to the number of knots.}

\item{var_est}{A character string; either "model", for model-based variance,
or "robust", to use the robust variance estimator.}

\item{var_est_type}{A character string; one of c("classic","DF","KC","MD","FG");
only relevant when var_est="robust".}

\item{return_ncs}{Logical; only relevant when exp_time="NCS". Specifies whether
the full covariance matrix for the calendar time parameters and the 
transformed treatment effect parameters are returned.}

\item{re_correlated}{Logical; specifies whether random treatment effect and random
intercept for cluster are correlated.}
}
\value{
A list of options
}
\description{
This should be used in conjunction with \code{\link{analyze}} to
    set parameters controlling analysis; see examples.
}
\examples{
dat <- load_data(time = "period", cluster_id = "cluster", individual_id = NULL,
treatment = "trt", outcome = "outcome_bin", data = sw_data_example)

analyze(dat = dat, method = "mixed", estimand_type = "TATE",  exp_time = "NCS",
family = binomial)

}
