% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot observed and predicted outcomes by cluster over time}
\usage{
plot_clusters(analysis_object, ncol = 3)
}
\arguments{
\item{analysis_object}{A list of class `sw_analysis`.}

\item{ncol}{Integer; number of columns in the faceted plot. Defaults to 3.}
}
\value{
A list with a `ggplot2` object of the actual and predicted outcomes by cluster.
}
\description{
Plot observed and predicted outcomes by cluster over time
}
\examples{
# Load data
test_data <- load_data(time ="period", cluster_id = "cluster", individual_id = NULL,
treatment = "trt", outcome = "outcome_cont", data = sw_data_example)

# Analyze using TATE estimand for exposure times 1 through 4
results_tate <- analyze(dat = test_data, method = "mixed", estimand_type = "TATE",
estimand_time = c(1, 4), exp_time = "ETI")

# Plot by cluster
plot_clusters(results_tate)

}
