% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot_design}
\alias{plot_design}
\title{Plot stepped wedge design}
\usage{
plot_design(dat)
}
\arguments{
\item{dat}{A dataframe containing the stepped wedge trial data.}
}
\value{
A list with a plot of the stepped wedge design.
}
\description{
Plot stepped wedge design
}
\examples{
# Load data
test_data <- load_data(time ="period", cluster_id = "cluster", individual_id = NULL,
treatment = "trt", outcome = "outcome_cont", data = sw_data_example)
# Plot design
plot_design(test_data)
}
