% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steppedwedge-package.R
\docType{package}
\name{steppedwedge-package}
\alias{steppedwedge}
\alias{steppedwedge-package}
\title{steppedwedge: An R Package for Analyzing Data from Stepped Wedge Cluster Randomized Trials}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This R package provides various functions and tools to help fit models for estimating treatment effects in stepped wedge cluster randomized trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://avi-kenny.github.io/steppedwedge/}
  \item Report bugs at \url{https://github.com/Avi-Kenny/steppedwedge/issues}
}

}
\author{
\strong{Maintainer}: Avi Kenny \email{avi.kenny@gmail.com}

Authors:
\itemize{
  \item David Arthur \email{david.arthur291@duke.edu}
  \item Yongdong Ouyang \email{yongdong.ouyang@sickkids.ca}
}

}
\keyword{internal}
